+++ tests/unit/core/data/conftest.py
@@ -7,7 +7,6 @@ from unittest.mock import MagicMock
 
 import numpy as np
 import pytest
-import torch
 from datumaro.components.annotation import Bbox, Label, Mask
 from datumaro.components.dataset import DatasetSubset
 from datumaro.components.dataset_base import DatasetItem
@@ -24,8 +23,6 @@ from otx.core.data.dataset.segmentation import (
     OTXSegmentationDataset,
     SegDataEntity,
 )
-from otx.core.data.entity.base import ImageInfo
-from otx.core.data.entity.segmentation import SegBatchDataEntity, SegBatchPredEntity
 from otx.core.data.mem_cache import MemCacheHandlerSingleton
 
 if TYPE_CHECKING:
@@ -61,27 +58,6 @@ def fxt_mock_dm_subset(mocker: MockerFixture, fxt_dm_item: DatasetItem) -> Magic
     mock_dm_subset.get.return_value = fxt_dm_item
     mock_dm_subset.__iter__.return_value = iter([fxt_dm_item])
     return mock_dm_subset
-
-
-@pytest.fixture(scope="session")
-def fxt_seg_data_entity() -> tuple[tuple, SegDataEntity, SegBatchDataEntity]:
-    img_size = (224, 224)
-    fake_image = torch.rand(img_size)
-    fake_image_info = ImageInfo(0, img_size[0], img_size[0],
-                           img_size[0], img_size[0])
-    fake_masks = torch.rand(img_size)
-    # define data entity
-    single_data_entity = SegDataEntity(fake_image,
-                                fake_image_info,
-                                fake_masks)
-    batch_data_entity = SegBatchDataEntity(1, [fake_image],
-                                           [fake_image_info],
-                                           [fake_masks])
-    batch_pred_data_entity = SegBatchPredEntity(1, [fake_image],
-                                           [fake_image_info], [],
-                                           [fake_masks])
-
-    return single_data_entity, batch_pred_data_entity, batch_data_entity
 
 
 @pytest.fixture(
